* Added tests * A random mac address will be assigned per reservation
